GCF-1053 Fix assertion `thd->wsrep_exec_mode == LOCAL_ROLLBACK`
This assertion triggers when thd is assigned a trx_id in between
rollback hooks. That is, before_rollback observes a thd without
a trx_id assigned yet, while after_rollback observes the same thd
with a trx_id assigned.
The fix here avoids such a situation by assigning a trx_id if
the thd does not have one yet in before_rollback hook.