submodules.cmake: git returns 128 if not in a repository
It is possible that a .git directory isn't definitive on the
existance of a working git tree. A git worktree over sshfs
for instance will fail to build unless the main repository
also exists in the same directory. Adding this extra test
will make the detection that little more ruggardised for these
odd build environments.

Signed-off-by: Daniel Black <daniel.black@au.ibm.com>