Instant ADD COLUMN: Revert changes to the adaptive hash index
The adaptive hash index is built on a prefix of the index keys,
on B-tree index leaf pages.

Instant ADD COLUMN does not change the format of secondary indexes:
the leaf page records will continue to contain
(secondary key columns,primary key columns).

The format of clustered index pages is changed by instant ADD COLUMN,
but even that change does not affect the format of the first few
columns and the record header. On the clustered index, the adaptive
hash index is always built on the PRIMARY KEY columns or a prefix of
them. The instantly added columns would be somewhere after the
DB_TRX_ID,DB_ROLL_PTR and any columns that were specified when the
table was last created or rebuilt. Thhe adaptive hash index can
safely access the PRIMARY KEY columns without caring about
instant ADD COLUMN.