GCF-1085 Re-enable MTR test galera_3nodes_sr.GCF-817
Test GCF-817 checks that the master of an SR transaction removes
SR table entries at right time, as it drops from the cluster and
then rejoins it.
The test was disabled as part of the GCF-990 refactoring, which
enforced transactions in 'QUERY_IDLE' state to be aborted by the
rollbacker thread. This was unexpected, as the test was checking
for entries to be present in SR table after going back to a
Primary view, and are no longer there.
Essentially, it was expected that the rollback would happen at
the subsequent interaction with the client, whereas now the
rollbacker takes care of rolling back the trx.
This patch adjusts the test accordingly.

There is also a minor change to wsrep_rollback_SR_connections()
that avoids sending a rollback fragment for SR transaction whose
master went non-Primary. The sending of the rollback fragment was
causing a warning for this test, and it was not needed at all as
slaves rollback all SR transactions that become non-Primary.