MDEV-13359 Enable ALTER TABLE...ALGORITHM=INPLACE for compressed columns
innobase_rec_reset(): Remove. This function was introduced in the
InnoDB Plugin for MySQL 5.1, which later evolved into MySQL 5.5.
There used to be a bug that ADD UNIQUE INDEX would not always correctly
report the duplicate key value of the secondary index. This function
ensured that instead of reporting total garbage values, InnoDB
would report NULL.

It looks like the function was made unnecessary in MySQL 5.6.6 by
https://github.com/mysql/mysql-server/commit/d143097eb1e02904baed34990db637ff68fe3113

The corresponding test was subsequently adjusted in
https://github.com/mysql/mysql-server/commit/fde80cf49d4b7635d2aef47d8a77681f2df62105

The ALTER TABLE tests were imported to MariaDB as part of MDEV-13625,
and these tests do pass with this change.

The unnecessary function did not do any harm before MDEV-11371 introduced
compressed columns.

One question remains: What if we needed to report a duplicate key value
for a compressed column? The simple answer is that the test
main.column_compression demonstrates that no indexes can be defined
on compressed columns.