Fix bogus rec_get_offsets() debug assertion failures for ROW_FORMAT=REDUNDANT
When the debug parameter 'bool leaf' was added to rec_get_offsets(),
also some debug assertions for reading the heap_no of ROW_FORMAT=REDUNDANT
records were added. However, the heap number is uninitialized when
offsets are being computed for to-be-inserted records.

For debug builds, initialize the heap number to a dummy value, so that
the record will be interpreted as 'user record'. The infimum and supremum
pseudo-records are never copied from the page frame and never inserted;
they are part of the page creation.

rec_convert_dtuple_to_rec_old(): Remove a bogus memset() in debug builds.