On second ADD COLUMN, update the 'default row' record
btr_cur_optimistic_update(), btr_cur_pessimistic_update(): Do not trim
the 'default row' record.

innobase_instant_add_try(): Compute and apply the update vector for
the 'default row'.

FIXME: The rollback of the 'default row' operations (that is, the
ROLLBACK of the ALTER TABLE…ADD COLUMN operation) has not been tested
and is unlikely to work. We need a test that flushes the redo log
and kills the server at the end of innobase_instant_add_try(), before
the ALTER TABLE transaction is committed.