Empty the PAGE_FREE list after deleting or shortening the 'default row'
We must know the size of each entry in the PAGE_FREE list.
After rolling back an instant ADD COLUMN operation, if the
'default row' is put into the PAGE_FREE list, we would be
unable to determine its size (if it is in ROW_FORMAT=COMPACT
or ROW_FORMAT=DYNAMIC) because it would contain more fields
than the rolled-back definition of the clustered index.