SQL: thd_start_utime() fix
Using start_utime is wrong, because it is my_interval_timer() which is
time since unspecified (but always the same) point in the past.

See comment for my_interval_timer():

The value is not anchored to any specific point in time (e.g. epoch) nor
is it subject to resetting or drifting by way of adjtime() or settimeofday(),
and thus it is *NOT* appropriate for getting the current timestamp. It can be
used for calculating time intervals, though.