Use cmake timestamp macro
Without this patch it is hard to build mariadb packages
that do not differ between builds.

See https://reproducible-builds.org/ for why this matters.

With this patch, it is possible, because recent cmake versions
honor the SOURCE_DATE_EPOCH environment variable.

Also allows building on platforms that are not UNIX or WIN32
as a side-effect.